vk_project/requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
Tarantool==1.1.0
PyJWT==2.9.0
requests==2.27.1
